Move logger under utils
